Do not fill in comboString when we are not using it.
According to callgrind, blobwars was spending 4% of its time in
enemyBulletCollisions(), calling snprintf() to fill in comboString. However,
this string was only used when an enemy was killed to check against objectives.
So now snprintf() has been moved to where it actually is being used, dropping
time spent in that function by 3.8%.