Remove bogus Exit LINEDEF in BioMech Communications.
This LINEDEF prevents the player from going into the small alcoves in the right
wall of the tank.  However, it is buggy, and in some cases the LINEDEF can be
triggered in such a way that the player cannot back out of it anymore.  After
removing, the bug is gone, and the player can enter the alcove a little bit,
but this is harmless.