Incorrect reference to SDL_FRMEWORK
There are two references to SDL_FRAMEWORK that should have been written
FRAMEWORK_SDL. In CGame.cpp, I think the section should be removed in it's
entirety as there is really no need for it and correcting the define results in
failure when compiling. In init.cpp, correcting the define allows Mac OS X
builds to assign a custom icon. If this is not fixed, the icon gets replaced
every time the game is run.