Make the window resizable, and set a reasonable initial size.
Now that SDL takes care of properly scaling the graphics, there is no
need for the window to be exactly 640x480. So allow the user to resize
it. To cope with high DPI screens, set the initial window size to a
multiple of 640x480 if it would fill less than half the width and
height.