Remove font size guessing hack.
According to the documentation, the point size given to TTF_OpenFont()
is always assuming 72 DPI, so there is no need to repeatedly trying to
load the font and checking the font metrics until we hit upon a point
size that matches the pixel size we want. Instead we can just give the
corresponding point size directly.