Redo all non-class header files.
The problem with the header files was that they contained forward
declarations for the corresponding .cpp file, instead of exporting the
declarations of the functions defined in the corresponding .cpp file
itself to other files. This caused a lot of duplication and potential
errors.

With some scripting, all functions that could be made static were found
and appropriately tagged. D. Richard Hipp's makeheaders tool was then
used to generate new .h files.